# Improve performance of map iteration using reflect.MapIter