This PR improves performance of iterating maps, using [reflect.MapIter](https://pkg.go.dev/reflect#MapIter) introduced in Go 1.12.
```
name       old time/op    new time/op    delta
Plain-12     2.21µs ± 2%    2.20µs ± 1%    ~     (p=0.548 n=5+5)
Logfmt-12    1.62µs ± 3%    1.60µs ± 0%  -1.16%  (p=0.016 n=5+4)
JSON-12      1.66µs ± 4%    1.60µs ± 1%  -3.68%  (p=0.008 n=5+5)
```